Move prerendering info Prerendering module
